+++ book.clj
@@ -730,10 +730,18 @@ v/table-viewer
     Moving --> Crash
     Crash --> [*]")
 
-;; ## 🙈 Controlling Visibility
+;; ## ⚙️ Customizations
 
-;; Visibility for code and results can be controlled document-wide and
-;; per top-level form. By default, Clerk will always show code and
+;; Clerk allows easy customization of visibility, result width and
+;; budget. All settings can be applied document-wide using `ns`
+;; metadata or a top-level settings marker and per form using
+;; metadata.
+
+;; Let's start with a concrete example to understand how this works.
+
+;; ### 🙈 Visibility
+
+;; By default, Clerk will always show code and
 ;; results for a notebook.
 
 ;; You can use a map of the following shape to set the visibility of
@@ -792,6 +800,44 @@ v/table-viewer
 ;;    (rand-int 42) ;; code will be visible
 ;;
 ;; This comes in quite handy for debugging too!
+
+;; ### 🍽 Table of Contents
+
+;; If you want a table of contents like the one in this document, set the `:nextjournal.clerk/toc` option.
+;;
+;;    (ns doc-with-table-of-contents
+;;      {:nextjournal.clerk/toc true})
+;;
+;; If you want it to be collapsed initially, use `:collapsed` as a value.
+
+;; ### 🔮 Result Expansion
+
+;; If you want to better see the shape of your data without needing to
+;; click and expand it first, set the
+;; `:nextjournal.clerk/auto-expand-results?` option.
+
+
+^{::clerk/visibility {:code :hide}
+  ::clerk/auto-expand-results? true}
+(def rows
+  (take 15 (repeatedly (fn []
+                         {:name (str
+                                 (rand-nth ["Oscar" "Karen" "Vlad" "Rebecca" "Conrad"]) " "
+                                 (rand-nth ["Miller" "Stasčnyk" "Ronin" "Meyer" "Black"]))
+                          :role (rand-nth [:admin :operator :manager :programmer :designer])
+                          :dice (shuffle (range 1 7))}))))
+
+;; This option might become the default in the future.
+
+
+;; ### 🙅🏼‍♂️ Viewer Budget
+
+;; In order to not send too much data to the browser, Clerk uses a per-result budget to limit. You can see this budget in action above. Use the `:nextjournal.clerk/budget` key to change its default value of `200` or disable it completely using `nil`.
+
+^{::clerk/budget nil
+  ::clerk/visibility {:code :hide}
+  ::clerk/auto-expand-results? true}
+rows
 
 ;; ## 🧱 Static Building